+++ Changelog.md
@@ -37,6 +37,7 @@ Ruby 2.0 is no longer officially supported.
 * Hide post title of limited post in comment notification email [#5843](https://github.com/diaspora/diaspora/pull/5843)
 * More and better environment checks in script/server [#5891](https://github.com/diaspora/diaspora/pull/5891)
 * Enable aspect sorting again [#5559](https://github.com/diaspora/diaspora/pull/5559)
+* Submit messages in conversations with Ctrl+Enter [#5910](https://github.com/diaspora/diaspora/pull/5910)
 
 # 0.5.0.0